backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
firebase-admin==5.0.3
python-dotenv==0.19.0
pyjwt==2.1.0 